[user-service] #42 move auth to separate controller
