+++ src/main/java/com/itmo/microservices/shop/user/api/model/AuthenticationResult.java
@@ -1,27 +1,15 @@
 package com.itmo.microservices.shop.user.api.model;
 
+import java.util.UUID;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
 public class AuthenticationResult {
+
     private String accessToken;
     private String refreshToken;
-
-    public AuthenticationResult(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-
-    public void setRefreshToken(String refreshToken) {
-        this.refreshToken = refreshToken;
-    }
+    private UUID uuid;
 }
