+++ src/main/kotlin/com/itmo/microservices/shop/common/security/SecurityConfiguration.kt
@@ -30,8 +30,8 @@ class SecurityConfiguration(
             .csrf().disable()
             .authorizeRequests()
             .antMatchers(HttpMethod.POST, "/users").permitAll()
-            .antMatchers(HttpMethod.POST, "/users/authentication").permitAll()
-            .antMatchers(HttpMethod.POST, "/users/refresh").hasAuthority("REFRESH")
+            .antMatchers(HttpMethod.POST, "/authentication").permitAll()
+            .antMatchers(HttpMethod.POST, "/authentication/refresh").hasAuthority("REFRESH")
             .antMatchers(HttpMethod.PUT, "/items/{id}").hasAuthority("ADMIN")
             .antMatchers(HttpMethod.POST, "/items").hasAuthority("ADMIN")
             .antMatchers(HttpMethod.DELETE, "/items/{id}").hasAuthority("ADMIN")
